update subscrible and share link
