+++ ProxySU_Core/Views/MainWindow.xaml.cs
@@ -1,9 +1,9 @@
 ﻿using MahApps.Metro.Controls.Dialogs;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
+using ProxySU_Core.Common;
 using ProxySU_Core.Models;
 using ProxySU_Core.ViewModels;
-using ProxySU_Core.ViewModels.Developers;
 using ProxySU_Core.Views;
 using Renci.SshNet;
 using System;
@@ -117,6 +117,36 @@ namespace ProxySU_Core
             Application.Current.Resources.MergedDictionaries[0] = resource;
         }
 
+        private void ExportXraySettings(object sender, RoutedEventArgs e)
+        {
+            StringBuilder sb = new StringBuilder();
+            foreach (var record in Records.Where(x => x.IsChecked))
+            {
+                record.Settings.Types.ForEach(type =>
+                {
+                    var link = ShareLink.Build(type, record.Settings);
+                    sb.AppendLine(link);
+                });
+            }
+            var tbx = new TextBoxWindow("分享链接", sb.ToString());
+            tbx.ShowDialog();
+        }
+
+        private void ExportXraySub(object sender, RoutedEventArgs e)
+        {
+            StringBuilder sb = new StringBuilder();
+            foreach (var record in Records.Where(x => x.IsChecked))
+            {
+                record.Settings.Types.ForEach(type =>
+                {
+                    var link = ShareLink.Build(type, record.Settings);
+                    sb.AppendLine(link);
+                });
+            }
+            var result = Base64.Encode(sb.ToString());
+            var tbx = new TextBoxWindow("订阅内容", result);
+            tbx.ShowDialog();
+        }
 
         private void AddHost(object sender, RoutedEventArgs e)
         {
