+++ ProxySU_Core/Views/TextBoxWindow.xaml
@@ -1,0 +1,17 @@
+﻿<metro:MetroWindow x:Class="ProxySU_Core.Views.TextBoxWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:ProxySU_Core.Views"
+        xmlns:metro="http://metro.mahapps.com/winfx/xaml/controls"
+        mc:Ignorable="d"
+        Title="TextBoxWindow" Height="450" Width="800">
+    <Grid>
+        <TextBox Margin="20"
+                 Text="{Binding Path=Message}"
+                 TextWrapping="Wrap"
+                 FontSize="14"
+                 Style="{StaticResource MahApps.Styles.TextBox}" />
+    </Grid>
+</metro:MetroWindow>
